Fix edit CRUD Book

+++ app/Views/authors/index.php
@@ -3,9 +3,9 @@
 
 <div>
     <div class="container mb-2">
-        <a href="/authors/create" class="btn btn-primary">
+        <?= (session()->get('user_type') === 'client') ? '' : '<a href="/authors/create" class="btn btn-primary">
             Add authors
-        </a>
+        </a>' ?>
     </div>
     <div class="container my-4 d-flex flex-row gap-2 align-items-center">
         <input placeholder="type it" class="input" name="text" type="text" />
@@ -19,7 +19,10 @@
                 <tr>
                     <th scope="col">id</th>
                     <th scope="col">Name</th>
-                    <th scope="col">Actions</th>
+                    <?php if (session()->get('user_type') === 'admin'): ?>
+                        <th scope="col">Actions</th>
+                    <?php endif; ?>
+
                 </tr>
             </thead>
             <tbody>
@@ -27,10 +30,12 @@
                     <tr>
                         <td><?= $author['id'] ?></td>
                         <td><?= $author['name'] ?></td>
-                        <td class="d-flex flex-row gap-3 text-center justify-center">
-                            <a href="/authors/edit/<?= $author['id'] ?>" class="btn btn-outline-warning" type="button">Edit</a>
-                            <a href="/authors/delete/<?= $author['id'] ?>" class="btn btn-outline-danger" type="button">Delete</a>
-                        </td>
+                        <?php if (session()->get('user_type') === 'admin'): ?>
+                            <td class="d-flex flex-row gap-3 text-center justify-center">
+                                <a href="/authors/edit/<?= $author['id'] ?>" class="btn btn-outline-warning" type="button">Edit</a>
+                                <a href="/authors/delete/<?= $author['id'] ?>" class="btn btn-outline-danger" type="button">Delete</a>
+                            </td>
+                        <?php endif; ?>
                     </tr>
                 <?php endforeach; ?>
